disk: add header file
